rk_ext: 3288 & 3288w only: add a path for libGLES_mali.so according to current SoC
Hardware revisions of mali_t760 in rk3288 and rk3288w are different,
so the mali_so's are different.
But the product_department requires that rk3288 and rk3288w must used a unified set of images.

With this commit, we can put the two mali_so's in a unified system_img,
in the path /system/vendor/lib/egl/rk3288/libGLES_mali.so and
./vendor/lib/egl/rk3288w/libGLES_mali.so,
and the image can support rk3288 and rk3288w both.

This only extends the process of searching libGLES_mali.so,
the default process has a higher priority.

Change-Id: Ib904e178e225a7848c5dc91378a1d58adad7137f
Signed-off-by: chenzhen <chenzhen@rock-chips.com>